LayoutTests/accessibility/help-text-expected.txt

Tests accessibility help text in a variety of examples.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS: Help text for <a> is ""
PASS: Help text for <a> is "bar"
PASS: Help text for <a> is "baz"
PASS: Help text for <a> is ""
PASS: Help text for <h1> is "bar"
PASS: Help text for <a> is ""
PASS: Help text for <button> is ""
PASS: Help text for <button> is "bar"
PASS: Help text for <button> is "baz"
PASS: Help text for <button> is ""
PASS: Help text for <table> is "table summary"
PASS: Help text for <input> is ""
PASS: Help text for <div> is ""
PASS: Help text for <main> is "bar"
PASS: Help text for <button> is ""
PASS: Help text for <main> is ""
PASS: Help text for <div> is ""
PASS: Help text for <a> is "bar"
PASS: Help text for <div> is ""
PASS successfullyParsed is true

TEST COMPLETE
